Move more fixtures
